refactor(backend): rename data-db to postgres-data-db
